requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub
spotipy
google-cloud-storage